requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
ccxt==4.0.78
ta==0.10.2
joblib==1.3.1
tqdm==4.65.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
pyyaml==6.0.1
